+++ src/MetaheuristicsPlatform/Algorithms/Matheuristics/FeasibilityPump/FeasibilityPumpMatheuristicParameters.cs
@@ -1,0 +1,17 @@
+using MetaheuristicsPlatform.Parameters;
+
+namespace MetaheuristicsPlatform.Algorithms.Matheuristics.FeasibilityPump;
+
+public sealed class FeasibilityPumpMatheuristicParameters : IMetaheuristicParameters
+{
+    public int MaximumIterations { get; init; } = 20;
+    public int NodeLimit { get; init; } = 1000;
+
+    public void Validate()
+    {
+        if (MaximumIterations <= 0)
+            throw new ArgumentOutOfRangeException(nameof(MaximumIterations));
+        if (NodeLimit <= 0)
+            throw new ArgumentOutOfRangeException(nameof(NodeLimit));
+    }
+}